server/requirements.txt
beautifulsoup4==4.12.2
feedparser==6.0.10
Flask==2.2.3
Flask_Cors==3.0.10
flask_sqlalchemy==3.0.3
openai==0.27.2
python-dotenv==1.0.0
python_dateutil==2.8.2
schedule==1.1.0
selenium==4.8.3
